refactor: 패키지 구조 재설계 및 클래스 위치 이동
