+++ app/src/main/java/woowacourse/shopping/repository/http/api/ProductApiService.kt
@@ -1,9 +1,11 @@
-package woowacourse.shopping.repository.http.product
+package woowacourse.shopping.repository.http.api
 
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
+import woowacourse.shopping.repository.http.dto.product.ProductPageResponseDto
+import woowacourse.shopping.repository.http.dto.product.ProductResponseDto
 
 interface ProductApiService {
     @GET("products")
