+++ app/src/main/java/woowacourse/shopping/repository/http/dto/product/ProductResponseDto.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.repository.http.product
+package woowacourse.shopping.repository.http.dto.product
 
 import kotlinx.serialization.Serializable
 import woowacourse.shopping.model.Money
