+++ app/src/test/java/woowacourse/shopping/repository/HttpProductRepositoryTest.kt
@@ -12,10 +12,10 @@ import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import woowacourse.shopping.repository.http.product.HttpProductRepository
-import woowacourse.shopping.repository.http.product.ProductNetworkException
-import woowacourse.shopping.repository.http.product.ProductParsingException
-import woowacourse.shopping.repository.http.product.ProductResponseException
+import woowacourse.shopping.repository.http.exception.product.ProductNetworkException
+import woowacourse.shopping.repository.http.exception.product.ProductParsingException
+import woowacourse.shopping.repository.http.exception.product.ProductResponseException
+import woowacourse.shopping.repository.http.repository.HttpProductRepository
 
 class HttpProductRepositoryTest {
     private lateinit var mockWebServer: MockWebServer
